requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.2
pydub==0.25.1
gunicorn==21.2.0
reportlab==4.0.4
python-docx==0.8.11
markdown==3.5.1
flask-cors==4.0.0